# backend/api.yaml
openapi: 3.0.0
paths:
...
  /api/v1/auth/register:
    post:
      description: >-
        Tạo tài khoản mới cho khách hàng với thông tin cá nhân và thông tin đăng
        nhập
      operationId: AuthController_register_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: Đăng ký thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Thông báo đăng ký thành công
                  user:
                    type: object
                    properties:
                      account_id:
                        type: number
                        description: ID tài khoản
                      username:
                        type: string
                        description: Tên đăng nhập
                      email:
                        type: string
                        description: Email
                      customer_id:
                        type: number
                        description: ID khách hàng
                      full_name:
                        type: string
                        description: Họ và tên
                      phone:
                        type: string
                        description: Số điện thoại
        '400':
          description: >-
            Yêu cầu không hợp lệ - Dữ liệu đầu vào không đúng định dạng hoặc
            thiếu thông tin bắt buộc
        '409':
          description: >-
            Xung đột dữ liệu - Tên đăng nhập, email hoặc số điện thoại đã tồn
            tại
      summary: Đăng ký tài khoản mới
      tags:
        - auth
  /api/v1/auth/login:
    post:
      description: >-
        Xác thực thông tin đăng nhập và trả về access token cùng với thông tin
        người dùng
      operationId: AuthController_login_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Đăng nhập thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT access token để xác thực các API khác
                  user:
                    type: object
                    properties:
                      account_id:
                        type: number
                        description: ID tài khoản
                      username:
                        type: string
                        description: Tên đăng nhập
                      email:
                        type: string
                        description: Email
                      phone:
                        type: string
                        description: Số điện thoại
                      role:
                        type: string
                        description: Vai trò của người dùng
                      full_name:
                        type: string
                        description: Họ và tên
        '400':
          description: Yêu cầu không hợp lệ - Thiếu thông tin đăng nhập
        '401':
          description: >-
            Thông tin đăng nhập không chính xác - Tên đăng nhập hoặc mật khẩu
            sai
        '404':
          description: >-
            Không tìm thấy tài khoản - Tài khoản không tồn tại hoặc đã bị vô
            hiệu hóa
      summary: Đăng nhập vào hệ thống
      tags:
        - auth
  /api/v1/auth/logout:
    post:
      description: Xóa refresh token và đăng xuất người dùng khỏi hệ thống
      operationId: AuthController_logout_v1
      parameters: []
      responses:
        '200':
          description: Đăng xuất thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Thông báo đăng xuất thành công
                    example: Đăng xuất thành công
        '401':
          description: Chưa xác thực - Token không hợp lệ hoặc đã hết hạn
      security:
        - JWT-auth: []
      summary: Đăng xuất khỏi hệ thống
      tags:
        - auth
  /api/v1/auth/profile:
    get:
      description: >-
        Trả về thông tin chi tiết của người dùng đang đăng nhập dựa trên access
        token
      operationId: AuthController_getProfile_v1
      parameters: []
      responses:
        '200':
          description: Thông tin profile của người dùng
          content:
            application/json:
              schema:
                type: object
                properties:
                  account_id:
                    type: number
                    description: ID tài khoản
                  username:
                    type: string
                    description: Tên đăng nhập
                  email:
                    type: string
                    description: Email
                  phone:
                    type: string
                    description: Số điện thoại
                  role:
                    type: string
                    description: Vai trò của người dùng
                  full_name:
                    type: string
                    description: Họ và tên
                  avatar:
                    type: string
                    description: URL ảnh đại diện
                    nullable: true
                  created_at:
                    type: string
                    format: date-time
                    description: Thời gian tạo tài khoản
        '401':
          description: Chưa xác thực - Token không hợp lệ hoặc đã hết hạn
        '404':
          description: Không tìm thấy thông tin người dùng
      security:
        - JWT-auth: []
      summary: Lấy thông tin profile của người dùng hiện tại
      tags:
        - auth
    patch:
      description: >-
        Cho phép người dùng cập nhật thông tin cá nhân như họ tên, email, số
        điện thoại, và thay đổi mật khẩu
      operationId: AuthController_updateProfile_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileDto'
      responses:
        '200':
          description: Cập nhật profile thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Thông báo cập nhật thành công
                  user:
                    type: object
                    properties:
                      account_id:
                        type: number
                        description: ID tài khoản
                      username:
                        type: string
                        description: Tên đăng nhập
                      email:
                        type: string
                        description: Email đã cập nhật
                      phone:
                        type: string
                        description: Số điện thoại đã cập nhật
                      full_name:
                        type: string
                        description: Họ và tên đã cập nhật
                      updated_at:
                        type: string
                        format: date-time
                        description: Thời gian cập nhật
        '400':
          description: Yêu cầu không hợp lệ - Dữ liệu đầu vào không đúng định dạng
        '401':
          description: Chưa xác thực - Token không hợp lệ hoặc đã hết hạn
        '404':
          description: Không tìm thấy thông tin người dùng
        '409':
          description: >-
            Xung đột dữ liệu - Email hoặc số điện thoại đã được sử dụng bởi tài
            khoản khác
      security:
        - JWT-auth: []
      summary: Cập nhật thông tin profile của người dùng hiện tại
      tags:
        - auth
  /api/v1/auth/refresh:
    post:
      description: Sử dụng refresh token để tạo access token mới khi token cũ đã hết hạn
      operationId: AuthController_refreshToken_v1
      parameters: []
      responses:
        '200':
          description: Làm mới token thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT access token mới để xác thực các API khác
        '401':
          description: >-
            Không thể làm mới token - Refresh token không hợp lệ, đã hết hạn
            hoặc không tồn tại
      summary: Làm mới access token
      tags:
        - auth
  /api/v1/auth/revoke:
    post:
      description: Thu hồi refresh token hiện tại, buộc người dùng phải đăng nhập lại
      operationId: AuthController_revokeToken_v1
      parameters: []
      responses:
        '200':
          description: Thu hồi token thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Thông báo thu hồi token thành công
                    example: Token đã được thu hồi thành công
        '401':
          description: Chưa xác thực - Token không hợp lệ hoặc đã hết hạn
      security:
        - JWT-auth: []
      summary: Thu hồi token
      tags:
        - auth
...
info:
  title: Cake POS API
  description: API documentation for Cake POS System - Hệ thống quản lý cửa hàng bánh ngọt
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
  schemas:
    RegisterDto:
      type: object
      properties: {}
    LoginDto:
      type: object
      properties: {}
    UpdateProfileDto:
      type: object
      properties:
        username:
          type: string
          description: Tên đăng nhập mới (tối thiểu 3 ký tự)
          example: newuser123
          minLength: 3
        password:
          type: string
          description: Mật khẩu mới (tối thiểu 8 ký tự)
          example: newpassword123
          minLength: 8
        last_name:
          type: string
          description: Họ
          example: Nguyễn
          maxLength: 70
        first_name:
          type: string
          description: Tên
          example: Văn A
          maxLength: 70
        gender:
          type: string
          description: Giới tính
          example: MALE
          enum:
            - MALE
            - FEMALE
            - OTHER
        phone:
          type: string
          description: Số điện thoại (10-11 số)
          example: '0901234567'
          maxLength: 15
        email:
          type: string
          description: Email
          example: user@example.com
          maxLength: 100
        position:
          type: string
          description: Chức vụ (chỉ dành cho nhân viên)
          example: Nhân viên bán hàng
          maxLength: 50
...